+++ dargon2_flutter_web/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 3.0.1
+- Add `initialize` method
+
 ## 3.0.0
 - Initial version (matched with [dargon2_flutter])
 - Made to split the mobile and web components of the plugin as mobile needs FFI while Web uses JS interop